fixed type of timer

+++ modern_durak_game_shared/modern_durak_game_shared.hxx
@@ -38,7 +38,7 @@
 #include <string>
 #include <variant>
 
-typedef std::vector<std::pair<std::string, long long int> > UserTimeMilliseconds;
+typedef std::vector<std::pair<std::string, uint64_t> > UserTimeMilliseconds;
 namespace shared_class
 {
 enum struct Move
